chore(add): add annotation, return types, TODO: check for return types

+++ backend/simpsonapi_playground/crud/actor.py
@@ -1,5 +1,6 @@
+from typing import Any, Literal, Tuple
 from sqlalchemy.orm import Session
-from sqlalchemy import asc, desc
+from sqlalchemy import ColumnElement, asc, desc
 
 from simpsonapi_playground.models.actor import Actor
 from simpsonapi_playground.models.character import Character
@@ -8,7 +9,7 @@ from simpsonapi_playground.schemas.actors_schemas import ActorCreate
 
 # TODO: Add safe delete
 # TODO: Admin role CRUD operations for Actor model
-def create_actor(db: Session, data: ActorCreate):
+def create_actor(db: Session, data: ActorCreate) -> Actor | None:
     new = Actor(**data.model_dump())
     db.add(new)
     db.commit()
@@ -21,18 +22,18 @@ def read_actors(
     limit: int,
     offset: int,
     sort: str,
-    order: str,
-):
+    order: Literal["asc", "desc"],
+) -> Tuple[list[Actor], int]:
     base_query = db.query(Actor)
     total = base_query.count()
 
-    sort_column_map = {
+    sort_column_map: dict[str, ColumnElement[Any]] = {
         "id": Actor.id,
         "first_name": Actor.first_name,
         "last_name": Actor.last_name,
         "cast": Actor.cast,
     }
-    sort_column = sort_column_map.get(sort, Actor.first_name)
+    sort_column: ColumnElement[Any] = sort_column_map.get(sort, Actor.first_name)
     if order == "asc":
         base_query = base_query.order_by(asc(sort_column), asc(Actor.id))
     else:
@@ -41,11 +42,11 @@ def read_actors(
     return (base_query.offset(offset).limit(limit).all(), total)
 
 
-def read_actor(db: Session, actor_id: int):
+def read_actor(db: Session, actor_id: int) -> Actor | None:
     return db.query(Actor).filter(Actor.id == actor_id).first()
 
 
-def update_actor(db: Session, actor_id: int, data: ActorCreate):
+def update_actor(db: Session, actor_id: int, data: ActorCreate) -> Actor | None:
     actor = db.query(Actor).filter(Actor.id == actor_id).first()
 
     if actor:
@@ -57,10 +58,12 @@ def update_actor(db: Session, actor_id: int, data: ActorCreate):
 
 
 # TODO: Admin role CRUD operations for Actor model
-def del_actor(db: Session, character_id: int):
-    actor = db.query(Actor).filter(Actor.id == character_id).first()
-    db.delete(actor)
-    db.commit()
+def del_actor(db: Session, actor_id: int) -> Actor | None:
+    actor = db.query(Actor).filter(Actor.id == actor_id).first()
+    if actor:
+        db.delete(actor)
+        db.commit()
+    return actor
 
 
 def get_characters_based_on_actor(
@@ -68,7 +71,7 @@ def get_characters_based_on_actor(
     actor_id: int,
     limit: int,
     offset: int,
-):
+) -> Tuple[list[Character], int]:
     base_query = db.query(Character).filter(Character.actor_id == actor_id)
 
     total = base_query.count()
@@ -78,5 +81,5 @@ def get_characters_based_on_actor(
     return items, total
 
 
-def get_actor_based_on_char(db: Session, character_id: int):
+def get_actor_based_on_char(db: Session, character_id: int) -> list[Actor]:
     return db.query(Actor).filter(Actor.id == character_id).all()
